fix: relocation of revoke func
